feat: create useBlockDelete hook
