+++ apps/journeys-admin/src/components/Editor/Slider/Content/Canvas/QuickControls/DeleteBlock/DeleteDialog/index.ts
@@ -1,0 +1,1 @@
+export { DeleteDialog } from './DeleteDialog'
